+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fix: Fix the stack overflow bug when Sequencex.flatten() is passed into TransformingSequence
 * new: Added 'tools4j-resources' module and 'tools4j-run' module
 * new: Sequencex add flatMapOfIterable, flatMapOfIterableTo, flatMapIndexed, flatMapIndexedOfIterable, flatMapIndexedOfIterableTo, flatMapIndexedTo method
+* new: Collectionx add flatMapIndexed, flatMapIndexedTo method
 
 ## v1.0.0-beta02
 improve: Improve code